Separate helpers from storybook config
